+++ src/utils/Oauth.ts
@@ -1,2 +1,2 @@
-export const getOauthLink = (clientId, redirectUri) =>
+export const getOauthLink = (clientId: string, redirectUri: string): string =>
     `https://github.com/login/oauth/authorize?scope=user&client_id=${clientId}&redirect_uri=${redirectUri}`;
